Change quoting for command invocation on Windows
Previously the command line for invoking maven was of the form
cmd /c "XXXX"
where "XXXX" was the actual mvn.cmd with options and hence (almost
always) would contain nested quotes which were escaped.
Recent changes to ProcessBuilder in the JVM (jdk8u232-b09,
jdk-11.0.5+10) tightened up validation for command line arguments
and it is now better to supply each element of the command line as
a separate string to ProcessBuilder and let the JDK handle escaping
the quotes.

See NETBEANS-3254 for details.